chore: moved styled comp to atoms
