+++ src/atoms/styled-image/styled.js
